[Backend] /tmp cleaner runs every 1000 seconds using celery-beat scheduler. Files written onto /tmp folder in the last 15 minutes wont be deleted. This time interval ensures that when a user/service is writing onto /tmp, the files are not deleted
(cherry picked from commit 613be585fb73110448d9adfa095b1d470d7d35c4)

+++ apps/filebrowser/src/filebrowser/tasks.py
@@ -131,40 +131,52 @@ def document_cleanup_task( **kwargs):
 import os
 import shutil
 import psutil
+from datetime import datetime, timedelta
 
 @app.task()
-def check_disk_usage_and_clean(**kwargs):
+def check_disk_usage_and_clean_task(**kwargs):
     task_id = kwargs.get("qquuid")
-    cleanup_threshold = kwargs.get('cleanup_threshold')
+    default_cleanup_threshold = 90
+    cleanup_threshold = kwargs.get('cleanup_threshold', default_cleanup_threshold)
+    username = kwargs.get("username", "celery_scheduler")
     now = timezone.now()
-
-    kwargs["username"] = kwargs["username"]
-    kwargs["task_name"] = "tmp_cleanup"
-    kwargs["task_id"] = task_id
-    kwargs["parameters"] = cleanup_threshold
-    kwargs["progress"] = "0%"
-    kwargs["task_start"] = now.strftime("%Y-%m-%dT%H:%M:%S")
+    kwargs = {
+        "username": username,
+        "task_name": "tmp_cleanup",
+        "task_id": task_id,
+        "parameters": cleanup_threshold,
+        "progress": "0%",
+        "task_start": now.strftime("%Y-%m-%dT%H:%M:%S")
+    }
 
     disk_usage = psutil.disk_usage('/')
     if disk_usage.percent >= int(cleanup_threshold):
-        print("Disk usage is above 90%, cleaning up /tmp directory...")
+        LOG.info(f"Disk usage is above {cleanup_threshold}%, cleaning up /tmp directory...")
         tmp_dir = '/tmp'
         for filename in os.listdir(tmp_dir):
             file_path = os.path.join(tmp_dir, filename)
             try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-                check_disk_usage_and_clean.update_state(task_id=task_id, state='SUCCESS', meta=kwargs)
-                print("/tmp directory cleaned.")
+                file_modified_time = datetime.fromtimestamp(os.path.getmtime(file_path))
+                # Make file_modified_time timezone-aware
+                file_modified_time = timezone.make_aware(file_modified_time, timezone.get_default_timezone())
+                if file_modified_time < now - timedelta(minutes=15):
+                    if os.path.isfile(file_path) or os.path.islink(file_path):
+                        os.unlink(file_path)
+                    elif os.path.isdir(file_path):
+                        shutil.rmtree(file_path)
+                    LOG.info(f"Deleted {file_path}")
+            except PermissionError:
+                LOG.warning(f"Permission denied: unable to delete {file_path}")
             except Exception as err:
-                check_disk_usage_and_clean.update_state(task_id=task_id, state='FAILURE', meta=kwargs)
+                check_disk_usage_and_clean_task.update_state(task_id=task_id, state='FAILURE', meta=kwargs)
                 raise Exception(f"Failed to delete {file_path}. Reason: {err}")
 
+        kwargs["progress"] = "100%"
+        check_disk_usage_and_clean_task.update_state(task_id=task_id, state='SUCCESS', meta=kwargs)
+        LOG.info("/tmp directory cleaned.")
 
     else:
         kwargs["progress"] = "100%"
-        check_disk_usage_and_clean.update_state(task_id=task_id, state='SUCCESS', meta=kwargs)
+        check_disk_usage_and_clean_task.update_state(task_id=task_id, state='SUCCESS', meta=kwargs)
         LOG.info(f"Disk usage is {disk_usage.percent}%, no need to clean up.")
